requirements.txt
Django==4.0.4
python-decouple==3.6
psycopg2-binary==2.9.3
asgiref==3.5.1
backports.zoneinfo==0.2.1
betterpath==0.2.2
click==8.1.3
Django==4.0.4
djangorestframework==3.13.1
l==0.11.0
Pillow==9.1.0
psycopg2-binary==2.9.3
python-decouple==3.6
pytz==2022.1
vcversioner==2.16.0.0
zope.interface==5.4.0
